feat: Implement user authentication and registration features
- Added Protected and RedirectAuthenticatedUser components for route protection.
- Created Register page for user registration with form handling and validation.
- Developed AuthService for handling authentication API calls (register, login, logout, etc.).
- Introduced CommentService and LikeService for managing comments and likes on posts.
- Implemented PostService for CRUD operations on posts.
- Enhanced Redux store with AuthSlice and PostSlice for managing authentication and post states.
- Added types for User and Post to improve type safety.
- Updated server-side controllers for user, post, comment, and like functionalities.
- Configured axios for API requests with base URL and credentials.
- Added necessary routes for user, post, and comment management.
- Included error handling and response structure improvements in API responses.
- Integrated Tailwind CSS for styling and improved UI components.

+++ server/src/controllers/bookmark.controller.js
@@ -1,23 +1,26 @@
+import mongoose from "mongoose";
 import { Bookmark } from "../models/bookmark.model.js";
 import { Post } from "../models/post.model.js";
-import { User } from "../models/user.model.js";
 
 const bookmarkPost = async (req, res) => {
   const { postId } = req.params;
 
   const post = await Post.findById(postId);
-  const user = await User.findById(req.user._id);
-  
+  if(!post){
+     return res.status(404).json({success:false, message: "post not found" });
+  }
+
   const postAlreadySaved = await Bookmark.findOne({ savedPost: postId });
   if (postAlreadySaved) {
-    await Bookmark.findByIdAndDelete(postAlreadySaved._id)
-    return res.status(200).json({message:"post removed from bookmark"})
+    await Bookmark.findByIdAndDelete(postAlreadySaved._id);
+    return res.status(200).json({success:true, message: "post removed from bookmark" });
   }
   await Bookmark.create({
-    savedPost: post._id,
-    savedBy: user._id,
+    savedPost: new mongoose.Types.ObjectId(post._id),
+    savedBy: new mongoose.Types.ObjectId(req.user._id),
   });
-  return res.status(200).json({message:"post saved in bookmark"})
+
+  return res.status(200).json({success:true, message: "post saved in bookmark" });
 };
 
 export { bookmarkPost };
